Fix profiling + jit bug
The profiler should report times using the shared runtime that the func
used, not the currently available shared runtime via
JITSharedRuntime::get (they may be different if it was released but the
Func being profiled is still around).

Also just always put the profiler in the shared runtime, instead of
depending on a target flag. For static builds it'll get dead stripped
just like tracing, memoization, etc.