Fix bug in Pipeline where it was possible to generate invalid code or source files because the function name to emit was allowed to be empty. This bug was found when trying to build the glsl app.
To fix this the name given to functions when no name is given is made to
match the JIT behaviour by factoring out that code from
Pipeline::compile_jit() into Pipeline::generate_function_name().

This bug could be exposed by calling the following methods with the
default arguments (fn_name as empty string).

- Func::compile_to_c()
- Func::compile_to_header()
- Func::compile_to_module()

or by calling the overloaded versions of the following methods where
in some versions an empty string was passed as the function name

- Func::compile_to_assembly()
- Func::compile_to_bitcode()
- Func::compile_to_object()