Add a NoRuntime target feature
Prevents inclusion of the runtime in AOT-compiled code.

Uses the same split as JIT does, so inlined math stuff is still included
in every module.

Declarations for the runtime API are provided to codegen by including a
runtime module that contains them as extern declarations (using an
attribute((used)) trick). This same runtime module also simplifies the
similar logic for JITting.

Also added a method to just compile a runtime, and a generator command-
line flag that does that (so any Generator can give you a copy of the
runtime for any target).