Make memoization zero copy.
Completes support for new_expr and delete_stmt on Allocate IR node.

Memoizing an output Func is no longer allowed as the zero-copy
approach does not work for this case and we decided to make it an
invariant that memoize does not copy. To fix this, introduce an
intermediate Func that is compute_root. The intermediat Func can be
memoized.

Fix memoize test to remove memoization of output Funcs.

This will likely be followed by refactoring the StorageFlattening
changes into a separate pass.


Former-commit-id: 935dad936914c904796d30f6065b80ee6e276971