Factor Memoization transforms out of Storage Flattening pass. The rewrite_memoized_allocations pass is heavily dependent on what storage_flattening generates so this is not as much of a structural win as one might like. It is however a less intrusive code change and makes it obvious the Memoization code changes don't affect anythign if memoization is not used.
Change cache key construction to not depend on the length of filter
and function names. This uses 12 bytes for a pointer plus a 32-bit
counter instead of 8 bytes plus the sum of the lenghts of the filter
name and memoized Func name. In practice this should be a win, but it
is a very small one.

Remove Closure changes as they are no longer needed. This should
remove the diffs to that part of the code relative to master.

Only run memoization passes if there is at least one memoized Func to
save a bit of compile time.


Former-commit-id: 5b691a81a10dc56c34ba25803ad27bdfa1a27919