Unify generator test executable names for both the Makefile and CMake build systems to
generator_exec_aot_test_*
generator_exec_jit_test_*

The intermediate generated object file names don't match between the
build systems but that doesn't matter because that's an internal
implementation detail of the build systems that any test framework
doesn't need to care about (i.e. it only needs to care about the
executables that need to be run to execute a test).


Former-commit-id: 7a404efb2a6127163b773b50eccd144ee9bd0c36