Synchronize write calls to ensure atomicity.
While writes to a file descriptor opened with O_APPEND are atomic, those opened for write do not appear to share that behavior.

This can result in a malformed event stream that cannot be reliably parsed. Wrapping this call in a lock removes this subtle requirement on the fd provided to the halide_set_trace_file API.


Former-commit-id: 804a027bb6a03e89f01389e3e6cd042d53cc54b4