Make profiler use malloc and free
Was using halide_malloc and halide_free with a NULL user context.

Unlike other uses of halide_malloc and halide_free, this one can
allocate in one thread and free in another. This tripped up some code
that was assuming this wouldn't happen. It's a bad assumption, but
halide_malloc is probably not appropriate here anyway.


Former-commit-id: 4d123e2d4c3550ce28ed7a04297c906618e50549