Change the name of the generator executables used to generate object code for the generator test cases under CMake.
The reason for doing this is the ``generator_`` prefix is confusing
(the test case executables start with the same prefix). We now follow
the Makefile build system convention where the generator executable has
a ``.generator`` suffix.

There are also a few tweaks to the comments to make things a little
clearer.


Former-commit-id: bd102f021a6f7860b46b08366e89067043afa264