DO NOT MERGE: lesson 10 form for bug repro case.
This should cause a crash with a Trace/BPT error when run with certain width parameters.

An original conjecture that it was triggered on width params that were not even multiples of 4 was insufficient. While all multiples of 4 do avoid the problem, the first crashing value is 7 and 15, 30, and 34 also do not hit the error.

Additionally, the issue goes away if I use repeat_edge instead of constant_exterior. It also goes away if I get rid of the vectorization in the schedule (taking away the parallel step causes a different misbehavior).