Revert "Add _fp16 user defined operator for C++11. The header file for this is"
This reverts commit 12cbb43b99fca0d33bd03b68eb2c5dfe4daaa1cb

The feature was a little controversial due to the header file hack and
the complexity that it introduces.

The main advantage of the _fp16 operator is that it calls the string
constructor which means the value is only rounded once (once by
float16_t). If a float or double literal is passed then the value is
rounded TWICE (once by the C++ compiler and once by float16_t). There
are examples where this causes a problem (e.g. "65455.999" and
"1025.49995") but I'm not sure if there enough of these cases to make
it worth worrying about. So revert for now...