Strengthen bounds inference to handle a new case
The bounds of

cast<uint16_t>(clamp(some_float_expr, 0.0f, 4096.0f))

were [0, 65535]

which surprised and annoyed a user.

This happened because Halide saw the bounds of the innermost thing as
some complex thing involving both 0/4096 and also some_float_expr. I.e.
it hadn't conservatively backed off to the clamp args, and then the
narrowing logic bailed on the nasty expressions, because it was only
designed to deal with constants.

This PR explicitly attempts to simplify the bounds of the innards
towards constants in that case, and simplifies some of the surrounding
code at the same time (I think it was written before Type::can_represent
existed).