Tracing now knows which pipeline it's in
Added a pipeline start and pipeline end event that top-level
realizations are children of. This lets you run a program that has many
Halide pipelines in it and trace it as one monolithic thing without
getting confusion between different funcs with the same name in the
different pipelines (e.g. inline reductions like 'sum')

Also allowed for 32-character Func names, up from the old limit of 16.

Also print stats of each Func that the trace sees, to make it easier to
do layout.