Remove an unnecessary cast after widening types in match_types()
An example is

a.type() == {code = Halide::Type::Float, bits = 32, width = 1}
b.type() == {code = Halide::Type::Float, bits = 32, width = 8}

After widening the types match so there is no need to add a cast
around either expression. An example application where this was observed
was when running the generator for the "bilateral_grid" app.