Add build system infrastructure for supporting "pure runtime tests"
Pure runtime tests will tests Halide's runtime and therefore link
against the runtime and not libHalide. The convention followed is
that a test source file that starts with the prefix ``runtime_`` is
treated as a pure runtime test.

The motivation for doing this is so it easy to test the recently added
halide_float16_t to float/double conversion functions.

The convention of adding a folder per test type is not followed here
because we may want "correctness pure runtime tests" and "error pure
runtime tests".