Change the second argument to create_buffer_t intrinsics from a byte size to an Expr where the Type itself is the meaningful information. This encodes more complete type information which is needed immediately in the JavaScript branch and will be used everywhere when we add complete type info to buffer_t.
I chose to replace the elem_size argument isntead of add a new one as
the information is redundant and this is a simpler code change. I
cannot think of a valid situation were elem_size should be allowed to
disagree with the byte size of the Type. In fact, I can see an
argument for elminating elem_size when the type info is added to
buffer_t.