Explicitly support the @llvm.x86.vcvtph2ps.128 and @llvm.x86.vcvtph2ps.256 intrinsics for convert float16 to single and double.
LLVM 3.7 would generate the correct code (using the above intrinsics)
when the target was x86 with F16C but on LLVM 3.6 codegen would hit an
error. Explicitly using the intrinsics fixes this.

Something isn't quite right though. When a vectorization width of 3 is
used the wrong results are computed so doing this has been guarded
against using an internal_assert.