Add ``roundingMode`` as a member of the Cast IR node and necessary function/method signature changes to support this. It's purpose is to state how an expression should be rounded when performing the cast. A variant of cast that doesn't take a rounding mode has been kept for API compatibility
This ``roundingMode`` only has meaning when casting to a floating point
type and where this floating point type has less precision than the input
type. The code doesn't emit errors that reflect this currently because
initially the rounding mode is only going to be used when casting to
float16_t and for other conversions (e.g. double -> float) I don't want
to emit errors/warnings which can't be fixed by the user.

Currently this ``roundingMode`` is not respected but future work will
add support for it when converting float/double to float16.