Add support for @llvm.x86.vcvtps2ph.128 and @llvm.x86.vcvtps2ph.256 intrinsics on x86-64 (with f16c) which is used to give a native implementation of converting to float (IEEE-754 2008 binary32) to float16 (IEEE-754 2008 binary16).
Only the RNE rounding mode is tested because the
float16_realize_downcast.cpp fails when using the other rounding modes.
It looks like there is a difference between the software implementation
(halide_float_to_float16_bits) when rounding 65536.0f and -65536.0f for
(RZ, RU and RD) and the result given by Intel's ``vcvtps2ph``. It is
most likely a mistake in our software implementation.