Fix the test cases in float16_t_realize_downcast (and add an additional test case) so that on x86 with f16c all the hardware rounding modes are tested and pass.
This means there is a bug in the software implementation of rounding
single and double to float16 which is why using it in the previously
mentioned test has been commented out.

I need to fix the test cases in
``test/correctness/runtime_float16_t_downcast.cpp `` and then fix
the implementation so that the software implementation can be used.