Prevent the ``float16_t_realize_downcast.cpp`` and ``float16_t_realize_upcast.cpp`` tests from actually testing anything on non x86 targets.
The tests will most likely fail on other targets because the target
specific codegen (e.g. CodeGen_MIPS) haven't been taught to use the
software implementation when it's needed. This really needs to be fixed
but I don't have access to the other targets right now and I don't
want to break the tests for those targets.