Fix bug when creating static library archives from halide generated code in the CMake build under Linux.
Previously the ``q`` flag was being used with ``ar`` which
just appends the object file to the archive without checking if it
was already present. This would lead to multiple versions of an object
file ending up in the archive. Then if an application linked against
this archive it would pick the first object file (i.e. the stale one)
rather than the latest version of the object file.

Now we use the ``r`` flag instead which is inserts object files into
the archive with **replacement**.


Former-commit-id: 2ef04e6ff7cb9635a614f602ecedb3bb51a0b7d3