[QTOOL - 839] - Create aligned loads and stores when possible
To generate aligned loads and store, the following stars have to align
be aligned.

1. The base address of the image has to be aligned. Done by having the user
specify a code gen option called BuffersAligned. Use it like so.
         target.set_cgoption(Target::BuffersAligned);
2. Set min of buffers/images in question to multiples of the vector size.
3. Set the stride of the vectorized dimension to be a multiple of the vector
factor. For this we have added to the Argument/Buffer/Image classes an extra
data member called stride_multiples that holds the highest value that perfectly
divides the stride. Users can use the set_stride_multiple methods of ImageParam
and Buffers to use this feature.

Change-Id: Ie85be8b4bf47baa813e7d56499044ed281b2bca5
Reviewed-on: https://review-hexagon.quicinc.com/35209
Tested-by: Anshu Das Gupta <adasgupt@quicinc.com>
Reviewed-by: Ron Lieberman <ronl@quicinc.com>
Reviewed-by: Dan Palermo <dpalermo@codeaurora.org>
Reviewed-by: Pranav Bhandarkar <pranavb@codeaurora.org>
Tested-by: Pranav Bhandarkar <pranavb@codeaurora.org>