[QTOOL - 1051] - Create the TargetTransformInfo pass before the AlwaysInliner Pass.
The always-inliner pass creates TargetTransformInfo with a default instance
of TargetIRAnalysis. Creating the TargetTransforInfo before that using
the TargetIRAnalysis provide by the TargetMachine allows all passes
that required TargetTransformInfo such as LoopStrengthReduce to get accurate
transform info. This helps in the generation of post-incs because the target
is able to tell LSR that post-incs is a legal addressing mode for certain types.

Change-Id: I9b4ab34210b2ea936ed2dd50b1563681ec70a4cd
Reviewed-on: https://review-hexagon.quicinc.com/35207
Reviewed-by: Ron Lieberman <ronl@quicinc.com>
Tested-by: Anshu Das Gupta <adasgupt@quicinc.com>
Reviewed-by: Dan Palermo <dpalermo@codeaurora.org>
Reviewed-by: Pranav Bhandarkar <pranavb@codeaurora.org>
Tested-by: Pranav Bhandarkar <pranavb@codeaurora.org>