Get most tests working
Lots of little fixes to get the tests running. The big change here is
that type coercion of int immediates is now handled by having
operatorfoo overloads that take ints. match_types no longer has a
special rule for intimms. So you add two Exprs, it always promotes to
the larger type, even if one of them happens to be an intimm.  For
example:

(some_u8 + 3).type() == UInt(8)
(some_u8 + Expr(3)).type() == Int(32)

Obviously this change needs to be discussed more widely. It was done
because now that IntImm can be lots of different types, I would need a
casting rule to resolve what to do if both expressions are IntImms. It
was starting to smell wrong.