Add support for OpenCL images and zero-copy buffers
When the Target::CLImages feature is enabled along with the
Target::OpenCL feature, the code generator and runtime will use the
OpenCL image2d_t type for 2D images and image2d_array_t for 3D images,
falling back to buffers for other dimensions. The main effect is that
loads and stores are replaced by read_image and write_image calls. If
Halide generates a kernel that uses a read/write image parameter, the
Target::OpenCL20 feature must be enabled to signify that the program
will be ran on a platform that has OpenCL 2.0 or later (which has
support for read/write images).

This also adds support for zero-copy buffers in the OpenCL runtime. If
halide_opencl_device_malloc is passed in a buffer_t whose host pointer
is already a cl_mem object, it will skip the malloc as it assumes it is
a buffer/image allocated using an OpenCL extension such as
CL_MEM_EXT_HOST_PTR_QCOM that allows an OpenCL device to use host memory
instead of copying data back and forth. Note that to avoid issues in
some OpenCL implementations, the check only occurs if the buffer_t's
_padding[0] is set to 1.