[QTOOL - 1920]- Use a narrower vmux when we can.
This patch narrows out vmux instructions when it can. Consider this
sum(x, y) += select(absdiff <= threshold, cast<unsigned short>(In(x + r.x, y + r.y)), 0);,
where absdiff, threshold and In are all byte sized. In the current form, if the above
is vectorized by the vector size, we'd be dealing with vector pairs. However,
if we transform the above to
sum(x, y) += cast<unsigned short> select(absdiff <= threshold, (In(x + r.x, y + r.y)), 0);,
we are now dealing with only a vector sized select or vmux.

Change-Id: I24c0e051181201e056d389a86801787e255062b1
Reviewed-on: https://review-hexagon.quicinc.com/38941
Reviewed-by: Dan Palermo <dpalermo@codeaurora.org>
Reviewed-by: Ron Lieberman <ronl@quicinc.com>
Tested-by: Anshu Das Gupta <adasgupt@quicinc.com>
Reviewed-by: Pranav Bhandarkar <pranavb@codeaurora.org>
Tested-by: Pranav Bhandarkar <pranavb@codeaurora.org>