Redo loop partitioning
We now partition at every level from the outermost in, and we also
partition GPU loops.

The previous loop partitioning method (only partition innermost loop)
fell on its face for things like convolutions - the loop over the kernel
was split up instead of the loop over the image.

Included a benchmark that tests a tiled convolution with a runtime size.
If you never hit the boundaries, the boundary conditions are all free.
They cost a moderate amount if you do hit the boundaries a bunch. On the
GPU they seem to be free all the time.


Former-commit-id: 48285fb9d01f726abb0b8f51b64e4132cb704e74