[CMake] Partially Fix #920
When rebuilding the runtime CMake was not aware that the source files
depended on the header files included by the sources. This fixes
this issue but unfortunately only for the makefile generator.

An issue is open for using Ninja as a generator but it doesn't look like
it's been touched in a while https://cmake.org/Bug/view.php?id=13234


Former-commit-id: ee9b316d5c8c06ddcd91df4899a92d5e9ac4e66e