Fix infinite loop in simplifier
We rewrite (a - b < c) -> a < (c + b). If a and b are constants we then
rewrite it back to the original. If this (a - b) doesn't constant-fold,
despite a and b being constant, then we're in a loop. Solved by only
doing the motion for constants that we know will fold.


Former-commit-id: b35c4de8725b3ba368bebbecfa2089e15d158bd6