[QTOOL - 3163] - Align all intermediate memory to vector width
This patch aligns all intermediate memory in two ways - memory
on the heap is always aligned to a 128 byte boundary. Memory on
the stack is aligned to native vector width i.e. 64 bytes in single
mode and 128 bytes in double mode. On the stack the alignment is
adjusted only if we are saving a vector type.

My number for Schedule 9 on toyCamera with this is 37.17 cycles/pixel
and for Schedule 10 is ~49 cycles/pixel. This patch restores the perf
of all other kernels.

Change-Id: I839406b5c2c7442559cbf7cd13d94ab07a98377e
Reviewed-on: https://review-hexagon.quicinc.com/47010
Reviewed-by: Dan Palermo <dpalermo@codeaurora.org>
Tested-by: Anshu Das Gupta <adasgupt@quicinc.com>
Reviewed-by: Ron Lieberman <ronl@quicinc.com>