[QTOOL - 3379] - Codegen loads with stride 2.
When we codegen vectors with stride == 2, we could load two adjacent vectors
and then use shufflevector to pick out the lanes we are interested in.

For e.g. to load every other half word we could do
v0 = load (x)
v1 = load (x+64) //single mode.
v2 = shuffle_vector v0, v1, <0, 2, 4, 6, ... >

Currently we do roughly the same as above, except if we are dealing with an
external buffer like an image, we shift the second vector left by 1 so as to
read beyond the end of the image. How is this possible? Let's say we were lo
64 bytes from memory location 'x' and we wanted only every other byte, then
accessing bytes x to x+126. But, if we choose two load v0 and v1 like above,
v0 = vector from x to x+63.
v1 = vector from x+64 to x+127. i.e we access an extra byte, which we eventu
discard using the shuffle vector, but nevertheless, we do load an extra byte
potentially accessing more memory than we may be allowed.
However, in the HVX case, for now, let's go the shuffle_vector route.


Former-commit-id: f78610a826fb939cbe9a47316cb8d0fe4c8e5d71