Align all intermediate memory to vector width
This patch aligns all intermediate memory in two ways - memory
on the heap is always aligned to a 128 byte boundary. Memory on
the stack is aligned to native vector width i.e. 64 bytes in single
mode and 128 bytes in double mode. On the stack the alignment is
adjusted only if we are saving a vector type.


Former-commit-id: 5babb5ab124b530a24ec477d6fcae4a8624660c0