Generate aligned loads in place of unaligned loads for all datatypes.
So far, we had been generating the vmem+vmem+valign sequence in place of a vmemu
only for byte-vectors. We now generalize this for all datatypes.

This patch, though, works only for well-formed ramp bases. Let's say we are trying
to load an unaligned half-word vector with this ramp(c1*64 + c2, 1, 32), where c2
is a constant and it doesn't really matter what c1 is. This is well formed because
64 is a mulitple of the lanes, so it is aligned in that sense. Assuming that the buffer
is aligned, it means we are exactly 'c2' half-words away from an aligned address.
So we will generate the following sequence:
           v0 = vmem(64*c1)      // load aligned vector
           v1 = vmem(64*c1 + 64) // load next aligned vector.
           v2 = valign(v1, v0, (c2*2))


Former-commit-id: d145c12b572f931e9cfd5b43c6c00dd05acc9545