Save and restore values over a loop backedge only if they are used in a load or store.
The loop store forwarding pass was sub-optimally moving indices to vector loads
and stores outside the loop by saving the previous iteration value of an index that
depended on the loop induction variable and reloading it in the next iteration. This
caused vector indices to be loaded from the stack thereby making it impossible for
modulusremainder analysis to be able to tell the alignment of vector loads and stores.
This was also causing an entire ramp to be stored on the stack and this was causing
camera_pipe to fail on hexagon.


Former-commit-id: e1bdc6c20a2a26b1916d94a3ab28c61531afcc4d