Change type info to more cleanly represent chains of pointers.
Add support for C++ L-value and r-value references. This is not well
tested as the Param<T> class cannot take a reference as its template
parameter. (Main use case is to allow buffer_t to be passed by
reference in the C++ interface.)

Add type mappings for all the basic types.

Remove const from value parameters in C prototypes.