Differentiate between pure and impure intrinsics and externs
Various places in the compiler need to know if they can move around,
coalesce, or duplicate call nodes, and they each seem to use a different
heuristic. This PR adds call types that explicitly tag call nodes as
pure. Other calls are assumed to impure.

Also added an is_intrinsic helper that I've been meaning to do for ages,
to simplify checking if a call is a particular intrinsic.

A lot of the noise in the PR is just reindenting code.