Use valign and vlalign in slice_vector in CodeGen_Hexagon.cpp.
This alongwith HexagonAlignLoads should theoretically obviate
the need for custom lowering of loads in CodeGen_Hexagon, but I noticed
slow downs in median and sobel when that custom lowering was removed. So,
I have put that back in and I will remove that after investigation. For now,
all test/hexagon/apps pass.