In slice_vector for CodeGen_Hexagon get the arguments of the vcombine directly.
When creating valign/vlalign  inside slice_vector, if we can tell that the double
vector coming in is the result of a vcombine, then just pull its arguments out
instead of calling .lo and .hi on the double vector. This was needed because
otherwise, sobel slowed down because a few vector copies got left around. By explicitly
CSEing ourselves, we are able to ensure better register allocation.


Former-commit-id: 72f5c47ca46670ef1df681b72623115cac24d97d