Incorporate some more review comments.
This patch uses reduce_modulo_expr to reason about alignment of the ramp base.
HexagonAlignLoads is changed to AlignLoads so as to not indicate that the pass
is specific to Hexagon. However, at this point it is run only for Hexagon(HVX).
This patch also adds a version of reduce_expr_modulo that takes a scope of ModulusRemainder
objects.


Former-commit-id: 7ccbe7428185b7b47bcb7c2e0aa7d25ddb1dd900