Fix tbaa bug by nuking stack reuse.
We promise LLVM that our allocations don't alias, and it optimizes loads
and stores appropriately. Stack slot reuse makes that a lie. The
simplest solution is to disable stack slot reuse - llvm should be able
to do this sort of thing anyway.

Note that the similar logic that fuses GPU shared allocations is fine -
those all come from a single allocate node (__shared), so the tbaa tags
are good. That approach - actually fusing Allocate nodes in a separate
pass - may be a way to ressurrect this feature in the future.