Augment GenGen options more
(1) add a -n flag option to allow (optionally) specifying an arbitrary
output file name; if unspecified, the filename continues to match the
function name.
(2) Tweak the -e option to add .o and .h header files to the list;
previously there was no way to suppress either of these, but for
certain build environments it is sometimes desirable to generate .o but
not .h (or vice versa).

(Note that the change to -e doesn’t affect any existing tests or apps
in Halide AFAICT, but has the potential to break external code that
uses it; in practice, -e was previously only specified when debugging,
so it’s unlikely anyone will be affected by this.)