Handle self-references by breaking the cycles
We use reference counting to track memory. This is a problem for update
definitions where a Func recursively calls itself. We used to deal with
this by reducing the reference count appropriately, but the self-
reference can now appear in other ways (e.g. the predicate of an RDom),
which makes this approach no longer work in certain deletion orders.

This PR instead allows Call nodes of call type Halide which do not
contain a pointer to a Function object. Or more precisely, they now have
a nullable pointer (an IntrusivePtr<FunctionContents>) instead of a non-
nullable one (a Function). The self-references in an update definition
have these set to null.

This makes our pipelines acyclic by construction, so reference counting
works. However this unfortunately means that the func field of a Call is
no longer a useful way to query aspects of the Function being called.
You need the environment. Made this explicit to avoid errors, and fixed
the resulting errors.