Add static library outputoption
This adds a new output option to gengen (and related APIs) that allows
the output to be in a static library form (-e static_library) —
typically, .a on unixy systems, .lib on Windows.

This totally ditches the hand-rolled static library code recently
added, in favor of the more robust code that already existed inside of
LLVM. (Unfortunately, LLVM’s code doesn’t allow for “files” from memory
streams, so we now must emit temporary files in some cases; this is an
acceptable price to pay.)

As a proof-of-workability, it converts the Generator tests in Makefile
to emit the standalone runtime in lib form rather than object form.

Of course, at present this is of somewhat limited use (since the only
output is wrapping a single .o in a .a, which doesn’t really add much),
but it will make emitting multiple builds with cpu-feature-detection
much more seamless.