Add file_make_temp() utility
The code in Module::compile() can be subject to race conditions when
building in parallel (e.g. if there are duplicate rules that attempt to
make the same file); add a utility to safely create a unique temporary
file and use it to avoid this problem.

(Note: Windows code believed to be correct but not yet compiled or
tested.)