Defer likely heuristic until loop partitioning
We have a heuristic in ScheduleFunctions that decides whether or not to
apply a likely on the base for the ShiftInwards tail strategy. This is
to trade off code size vs performance. It was pretty broken in a bunch
of cases. The basic problem is that you can't tell at that point what's
going to be an innermost loop. This PR defers the decision until loop
partitioning time, which is after vectorization and unrolling. This
unfortunately required a new variant of likely.

likely always partitions a loop

likely_if_innermost converts to likely if it's in an innermost loop
(after vectorization and unrolling). Otherwise it just goes away.

Note that this doesn't necessarily mean we'll partition the innermost
loop. This would be bad, e.g. if it's a small loop over a convolution
kernel inside a large loop over pixels. We still partition the outermost
loop we can that simplifies the likely intrinsics. We just only care
about simplifying likely_if_innermost intrinsics if we find them in an
innermost loop.