Upgrade compile_multitarget() to output 'extra' files when requested
e.g., stmt files and suchlike. This is a little awkward (since multiple
files are output for these vs. a single output for .h and .a) but
probably the least weird solution here.

Note that .o files are still forbidden here: it’s not really useful and
potentially wrong without extra effort.