Upgrade compile_multitarget() to output 'extra' files when requested
e.g., stmt files and suchlike. This is a little awkward (since multiple
files are output for these vs. a single output for .h and .a) but
probably the least weird solution here.

Note that .o files are still forbidden here: it’s not really useful and
potentially wrong without extra effort.


Former-commit-id: 08a9c1d8b7b70e3cc098f388e0e56238c35e68b6