Got hvx thread pool working
Uses the ability to dynamically change the number of working threads in
the thread pool to cope with HVX128/HVX64 changes or even nested
parallelism that mixes HVX modes (why would you do that?).

Added a test for dynamically changing the number of working threads in
the thread pool. Didn't require any new thought in the thread pool -
turns out it was already possible. I just did some code cleanup in
there.


Former-commit-id: eac9c4403f6a1a4a8cbcff2d4570c2cf32581c8c