Add a target feature that fuzzes floating point stores
When the FuzzFloatStores target feature is on, the LSB of all floating
point stores is set to zero before storing.

Because we apply (and tell LLVM to apply) -ffast-math style floating
point optimizations in a target-specific way, you can't rely on Halide
code producing exactly the same outputs on every platform for code that
uses floats. This new target features is intended to flush out tests
that make this bad assumption. If a test passes with FuzzFloatStores
off, but fails with it on, it's likely to also fail on different
architectures.

There are probably all sorts of more elaborate ways to fuzz floating
point computations, but we've found that this surprisingly simple method
works well for finding the bad tests and has few false positives.


Former-commit-id: 3fc91b5cfd49efea609357a7ff34983b8d01a7eb