Fix Param<const-pointer> for C++ name mangling
We weren’t consistent about distinguishing const and non-const pointers
in the C++ name mangling code. Now we ensure that constness is always
preserved, which is necessary for C++ name mangling to properly match
between the .h and .o/.a.

(Note that user_context was formerly generated as const void*; now it
is a plain void*, which might require minor downstream casting changes.)


Former-commit-id: e6824a9aef68409374497ea3df8ff27234d5a3c9