Rework EliminateBoolVectors
... so that the intermediate IR is well-formed and produces the same
results on every platform. Now on Hexagon and OpenCL we promote bool
vectors to multi-bit "masks" immediately after creating them (with a
comparison). This is codegenned as a no-op, because that's what the
comparisons do natively. All of the bits of a mask must be zero when it
represents a false value. True values are represented by architecture-
specific values (0x01 in every byte on hexagon, 0xff in every byte on
opencl). There is a cast and a select equivalent that work on these
masks. They must be implemented per-backend in a way that works with
that backend's representation. Logical operators are handled with
bitwise operators.

Storing a mask, or casting it to an actual int, forces it to be zero or
one, because at that point it might escape one architecture and leak to
another (e.g. storing on Hexagon followed by loading on ARM).