Split Generator builds into libs
This reworks the Generator build rules so that instead of building a
binary directly from source, we create intermediate libraries that are
then linked into binaries.

As-is, this change is a no-op, but paves the way for some other
upcoming changes (which invole combining multiple precompiled
Generators into a single binary, then accessing them via the
GeneratorRegistry); doing this as a separate PR should make for a much
cleaner review.

(Note that we use shared libraries rather than static libraries, mainly
because static libraries will get dead-stripped away due to the way
Generators are registered, and there doesn’t seem to be a good,
portable “alwayslink” option across all systems.)