Avoid recompiling shaders when possible
My recent change (fixed by @abadams) to allow multiple GLSL stages made
it so that the runtime recompiles shaders even when not necessary.
This change checks (by kernel name) to see if the kernel(s) have
already been compiled and avoids the overhead of recompilation.

We _might_ want to do something more sophisticated for matching
(perhaps hashing the kernel text or something).