Remove register_metadata feature
This relied on nonstandard linker behavior (keeping unreferenced static
data that has side-effects in its ctors) that doesn’t work in all
situations; in particular, multitarget breaks it. It was put in place
for one client (Safelight) that can actually get by without it. So
let’s simplify the world and just remove it.


Former-commit-id: d127b2d0f32faee3056e92bdc127be3fc8ab55f9