Rework CMake Generator Rules
Existing CMake rules for Generator build/test were a mess of
unmaintainability. Major change is to move from a
glob-with-special-cases (Make-style) to an explicit-list-of-build-rules
(Bazel-style), which makes for a much saner setup. Also broke the
megalithic rule into smaller pieces to deal with CMake idiosyncrasies.
Desired goal is that the tests in test/CMakeLists.txt should be much
more sensible to read (and maintain) going forward.