Remove / privatize some Generator APIs
These were public calls used only internally to Google (AFAICT) for
implementing a multi-cpu-selection code; with the advent of a
Halide-provided feature for this, they are no longer needed and better
off gone.

Also enforce that set_generator_param_values() can be called at most
once per instance, which was always the intent.


Former-commit-id: 0bb313af87e18cc8775ea6a9fbe9995405edbb4a