Revise MSAN wrappers
Initial commit wasn’t able to handle buffers that were optimized away
(e.g. by VectorizeLoops); it also could fail if a buffer was elided
from a Closure but the assertion from the annotate call was not. This
PR:

(1) changes the runtime calls to return void, and skip assertions
(runtime failures were never expected anyway), and
(2) explicitly add calls to mark output buffers, and also to mark
buffers that are passed to extern functions.