Propagate MSAN to Hexagon targets
If the host is compiling with MSAN, the hexagon target needs to as
well… otherwise, we can end up injecting both the msan and msan_stubs
runtimes, and the weak linkage can pick the ‘wrong’ msan code, with the
result that an msan build uses non-msan-enabled stubs.

Reviewers: are there other places we need to propagate features across
targets in this way?

(IMHO this is an indictment of the msan-stubs approach; I’ll probably
follow up with another patch to try to eliminate the need for it)