allow compile_to_static_library() and compile_to_file() to use full paths
They assume that the filename_prefix is also the function name, so if
you pass in (say) /path/to/foo you’ll get a bad function name.

Fix to allow existing code to work is to just split the filename_prefix
on “/“ and take the final segment as the fn_name; this allows all
existing use cases to work seamlessly.