make OpenGL and RenderScript tests less env-var dependent
test/opengl and test/renderscript requires the build system to set
HL_TARGET and/or HL_JIT_TARGET correctly; it’s simpler and more robust
to set the Target features we need inside the tests, and not require
the build/test harness to do it. (Makefile did set them correctly, but
CMake never did.)

Also loosen error on sum_reduction a bit.


Former-commit-id: 5290aa9e5c8c8ba3306c222b6f8e5fde84da9a71