Simplify overly-clever Stub registration code
I tried to be clever and make an enhanced RegisterGeneratorAndStub
class that would fail at linktime (rather than runtime) if the
necessary Generator wasn’t present; this was overly complicated and (as
it turns out) error prone, since multiple instantiations of the same
templated class may not behave as I expected. Switch back to the
simpler-to-understand runtime-failing code.