Rework Stubname and RegisteredName to be the same for Generators
This is a bit of an experiment, but I think this may simplify our
nomenclature in the long run with Stubs. (Please don’t land this PR
without some discussion.)

The idea here is that we remove the separate “stubname” part of
HALIDE_REGISTER_GENERATOR, and use the existing registration name for
it, with a little smarts:

— if the registration name contains at least one “::”, it’s assumed to
be a fully-qualified namespace-and-classname that should be used to
name the generated Stub.

— if the registration name does not contain at least one “::”, it’s
assumed that the Generator doesn’t want a Stub (mostly, a legacy
Generator from before Stubs, though new Generators could also do this
if they know they never need useful Stubs generated).

Note also that the emitter now will always emit a .stub.h file for
every Generator (even for “legacy” Generators for which this isn’t
possibly); for older Generators, it’s basically just a file with a
comment saying “There is not Stub for this Generator”. This turns out
to simplify build rules for Bazel considerably since it is much
stricter about specifying which files are generated or not.


Former-commit-id: 9225d6b0afa9246a0a9fbd48314eabdbe5cac480