Revise fft_generator input/output images
With this change, inputs and outputs are now uniformly 3-dimensional,
with the order (x, y, c): real buffers have extent[c]=1, complex
buffers have extent[c]=2.

This is (to some extent) more a style issue than anything, but:
— having inputs or outputs that vary number-of-dimensions is
legal-but-weird; we’d prefer to have our apps present examples that we
consider to be non-weird.
— putting anything before (x,y) in the argument order is also
legal-but-weird; reordering to (x,y,c) allows for a more-customary
interleaved layout. (Note that the *memory* layout for complex buffers
is identical to what it was before, but the *order* of the dimensions
is now more customary.)