Update IntegerDivisionTable
find_inverse.c got renamed to .cpp (since it was C++ already), and now
generates matching .h and .cpp files. The signed tables were previously
generating 257 entries (but were only declared as 256); trimmed off the
extras. Also declared data as const because, well, if this mutates,
we’re in trouble.


Former-commit-id: 532ce20948205931459adfa3a809f6f35021250d