Buffer enhancements
Several enhancements to Halide::Buffer motivated by real-world usage.

Most notably there's a for_each_value that iterates over the values
without tracking coordinates. This produces very efficient inner loops,
and copy and fill are now implemented using this.

Also added a method for testing if a coordinate is contained within the
region of a buffer, added a method to cast a Buffer reference to a
buffer reference of another type, and added a static method to create a
buffer with the same shape as another buffer.


Former-commit-id: f13789e0cfbe9aa756c3a189be23b53fa6262b0d